# service/api/openapi/core-api-v1.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Statevia Core API",
    "description": "Core-API HTTP 契約。運用・SSE・冪等・IO-14 等の叙述は docs/core-api-interface.md を参照。",
    "version": "v1"
  },
  "paths": {
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Auth"
        ],
        "summary": "POST /v1/auth/login — パスワードログイン。",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "テナント ID。未指定時は \"default\"。",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/me": {
      "get": {
        "tags": [
          "Auth"
        ],
        "summary": "GET /v1/auth/me — 認証済み Principal 情報。",
        "parameters": [
          {
            "name": "X-Tenant-Id",
            "in": "header",
            "description": "テナント ID。未指定時は \"default\"。",
            "schema": {
              "type": "string",
              "default": "default"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthMeResponse"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "description": "機械可読なエラーコード。",
            "nullable": true
          },
          "message": {
            "type": "string",
            "description": "人が読む説明メッセージ。",
            "nullable": true
          },
          "details": {
            "description": "追加の構造化詳細（任意）。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "契約の error オブジェクト。"
      },
      "AuthMeResponse": {
        "type": "object",
        "properties": {
          "tenantId": {
            "type": "string",
            "description": "テナント内部 ID。",
            "format": "uuid"
          },
          "tenantKey": {
            "type": "string",
            "description": "外部キー。",
            "nullable": true
          },
          "principalId": {
            "type": "string",
            "description": "Principal ID。",
            "format": "uuid"
          },
          "email": {
            "type": "string",
            "description": "メールアドレス。",
            "nullable": true
          },
          "isTenantAdmin": {
            "type": "boolean",
            "description": "テナント管理者か。"
          }
        },
        "additionalProperties": false,
        "description": "認証済み Principal 情報。"
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiError"
          }
        },
        "additionalProperties": false,
        "description": "データ連携契約に基づく HTTP エラーレスポンス。{ \"error\": { \"code\", \"message\", \"details\" } }。"
      },
...
      "LoginRequest": {
        "type": "object",
        "properties": {
          "tenantKey": {
            "type": "string",
            "description": "外部向けテナントキー。",
            "nullable": true
          },
          "email": {
            "type": "string",
            "description": "メールアドレス。",
            "nullable": true
          },
          "password": {
            "type": "string",
            "description": "平文パスワード。",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "ログイン要求。"
      },
      "LoginResponse": {
        "type": "object",
        "properties": {
          "accessToken": {
            "type": "string",
            "description": "JWT アクセストークン。",
            "nullable": true
          },
          "expiresAt": {
            "type": "string",
            "description": "有効期限（UTC）。",
            "format": "date-time"
          },
          "tenantId": {
            "type": "string",
            "description": "テナント内部 ID。",
            "format": "uuid"
          },
          "tenantKey": {
            "type": "string",
            "description": "外部キー。",
            "nullable": true
          },
          "principalId": {
            "type": "string",
            "description": "Principal ID。",
            "format": "uuid"
          }
        },
        "additionalProperties": false,
        "description": "ログイン成功応答。"
      },
...
    }
  },
...
}